Fix an issue of not updating the OPEN key counters, after Addin is Removed
Fix an issue of not updating the OPEN key counters, after Addin is Removed. In this implementation, after removing the desired Key, we are collecting all the others, then deleting them. So in next loop at line 203, will insert them back, with correct OPEN key counters and preserving the original order. Feel free to enhance it further...